backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-decouple==3.8
firebase-admin==7.1.0
oss2==2.18.4
requests==2.31.0
openai==1.3.7 